requirements.txt
auraloss==0.4.0
hydra-core==1.3.2
matplotlib==3.10.0
nnAudio==0.3.2
numba==0.58.1
numpy==1.26.1
pandas==2.2.3
Pillow==10.1.0
pyaml==23.9.7
PyQt6==6.9.1
scipy==1.11.3
seaborn==0.13.2
soundfile==0.12.1
tensorboard==2.15.1
termcolor==2.4.0
threadpoolctl==3.2.0
torch==2.2.1
torchaudio==2.2.1
torchinfo==1.8.0
torchmetrics==1.2.0
torchvision==0.17.1
tqdm @ file:///C:/b/abs_f76j9hg7pv/croot/tqdm_1679561871187/work
